Se ajustarons los metodos para la generacion de reportes

+++ AppGia/Controllers/ReportesController.cs
@@ -10,7 +10,7 @@ namespace AppGia.Controllers
     {
         private ReportesHelper _reportesHelper=new ReportesHelper();
         
-        // POST: api/ProformaExcel
+     
         [HttpPost("generar")]
         public Dictionary<string,string> export([FromBody] ReportesRequest request)
 
@@ -28,7 +28,7 @@ namespace AppGia.Controllers
         }
         
         [HttpGet("parametros/{idReport}")]
-        public List<Dictionary<string, object>> getParametros(int idReport)
+        public List<Dictionary<string, object>> getParametros(Int64 idReport)
         {
             return _reportesHelper.getParametrosOf(idReport);
         }
